refactor: result 패키지 이동
