+++ src/main/java/chess/service/ChessService.java
@@ -10,14 +10,14 @@ import chess.domain.chessboard.ChessBoardFactory;
 import chess.domain.chesspiece.ChessPiece;
 import chess.domain.chesspiece.Color;
 import chess.domain.position.Position;
+import chess.domain.result.EndResult;
+import chess.domain.result.MoveResult;
+import chess.domain.result.StartResult;
 import chess.dto.ChessPieceDto;
 import chess.dto.ChessPieceMapper;
 import chess.dto.CurrentTurnDto;
 import chess.dto.MoveRequestDto;
 import chess.dto.RoomStatusDto;
-import chess.result.EndResult;
-import chess.result.MoveResult;
-import chess.result.StartResult;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
